+++ Sources/SwiftProtobuf/TextFormatScanner.swift
@@ -496,7 +496,7 @@ internal struct TextFormatScanner {
                 throw TextFormatDecodingError.malformedNumber
             }
             let n = try nextUInt()
-            let Int64MaxPlusOne: UInt = 0x8000000000000000
+            let Int64MaxPlusOne: UInt64 = 0x8000000000000000
             if n >= Int64MaxPlusOne { // -Int64.min
                 if n > Int64MaxPlusOne {
                     // Too large negative number
